requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
WTForms==2.3.3
APScheduler==3.7.0
python-dotenv==0.19.0
email_validator==1.3.0
gunicorn==21.2.0
